+++ aletheia_core/lib/aletheia_core_model.dart
@@ -18,6 +18,7 @@ export 'src/model/bible/verse/view/view.cross.reference.dart';
 export 'src/model/bible/verse/view/view.search.verse.dart';
 export 'src/model/bible/verse/view/view.verse.dart';
 export 'src/model/bible/verse/footnote.dart';
+export 'src/model/bible/verse/verse.audio.dart';
 //COMMENTARY
 export 'src/model/commentary/commentary.content.dart';
 export 'src/model/commentary/commentary.dart';