+++ aletheia_core/lib/src/model/bible/verse/verse.audio.dart
@@ -1,0 +1,9 @@
+/// Represents audio data for a Bible verse
+class VerseAudio {
+  List<int> audioBytes;
+  String audioFormat;
+  VerseAudio({
+    required this.audioBytes,
+    required this.audioFormat,
+  });
+}